Return an error on config upgrade command
And split the config commands over multiple files, following the
pattern we also use for other commands.
